小修改！

+++ Binatone-demo/Binatone-demo/libs/Binatone.framework/Headers/BinatoneOriginalData.h
@@ -17,10 +17,10 @@
 @end
 
 @interface BinatoneHistoryDataRecord : NSObject
-@property (nonatomic,assign) UInt8 br;//呼吸率
-@property (nonatomic,assign) UInt8 hr;//心率
-@property (nonatomic,assign) UInt8 status;//状态
-@property (nonatomic,assign) UInt8 statusValue;//状态值
+@property (nonatomic,assign) UInt32 br;//呼吸率
+@property (nonatomic,assign) UInt32 hr;//心率
+@property (nonatomic,assign) UInt32 status;//状态
+@property (nonatomic,assign) UInt32 statusValue;//状态值
 
 + (NSInteger)length;
